rust/direction: move direction to own file (cleanup)
Move the implementation of Direction to its own file, direction.rs.
