+++ rust/src/core.rs
@@ -19,7 +19,6 @@
 
 use std;
 use crate::filecontainer::*;
-use crate::debug_validate_fail;
 use crate::flow::Flow;
 
 /// Opaque C types.
@@ -41,70 +40,6 @@ pub const STREAM_TOCLIENT: u8 = 0x08;
 pub const STREAM_GAP:      u8 = 0x10;
 pub const STREAM_DEPTH:    u8 = 0x20;
 pub const STREAM_MIDSTREAM:u8 = 0x40;
-pub const DIR_BOTH:        u8 = 0b0000_1100;
-const DIR_TOSERVER:        u8 = 0b0000_0100;
-const DIR_TOCLIENT:        u8 = 0b0000_1000;
-
-#[repr(C)]
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum Direction {
-    ToServer = 0x04,
-    ToClient = 0x08,
-}
-
-impl Direction {
-    /// Return true if the direction is to server.
-    pub fn is_to_server(&self) -> bool {
-	matches!(self, Self::ToServer)
-    }
-
-    /// Return true if the direction is to client.
-    pub fn is_to_client(&self) -> bool {
-	matches!(self, Self::ToClient)
-    }
-
-    pub fn index(&self) -> usize {
-        match self {
-            Self::ToClient => 0,
-            _ => 1,
-        }
-    }
-}
-
-impl Default for Direction {
-    fn default() -> Self { Direction::ToServer }
-}
-
-impl std::fmt::Display for Direction {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::ToServer => write!(f, "toserver"),
-            Self::ToClient => write!(f, "toclient"),
-        }
-    }
-}
-
-impl From<u8> for Direction {
-    fn from(d: u8) -> Self {
-        if d & (DIR_TOSERVER | DIR_TOCLIENT) == (DIR_TOSERVER | DIR_TOCLIENT) {
-            debug_validate_fail!("Both directions are set");
-            Direction::ToServer
-        } else if d & DIR_TOSERVER != 0 {
-            Direction::ToServer
-        } else if d & DIR_TOCLIENT != 0 {
-            Direction::ToClient
-        } else {
-            debug_validate_fail!("Unknown direction!!");
-            Direction::ToServer
-        }
-    }
-}
-
-impl From<Direction> for u8 {
-    fn from(d: Direction) -> u8 {
-        d as u8
-    }
-}
 
 // Application layer protocol identifiers (app-layer-protos.h)
 pub type AppProto = u16;
@@ -297,19 +232,5 @@ pub fn sc_app_layer_decoder_events_free_events(
         if let Some(c) = SC {
             (c.AppLayerDecoderEventsFreeEvents)(events);
         }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_direction() {
-	assert!(Direction::ToServer.is_to_server());
-	assert!(!Direction::ToServer.is_to_client());
-
-	assert!(Direction::ToClient.is_to_client());
-	assert!(!Direction::ToClient.is_to_server());
     }
 }
