+++ rust/src/lib.rs
@@ -140,6 +140,7 @@ pub mod feature;
 pub mod sdp;
 pub mod ldap;
 pub mod flow;
+pub mod direction;
 
 #[allow(unused_imports)]
 pub use suricata_lua_sys;
