+++ rust/src/smb/smb1.rs
@@ -19,8 +19,7 @@
  * - check all parsers for calls on non-SUCCESS status
  */
 
-use crate::core::*;
-
+use crate::direction::Direction;
 use crate::smb::smb::*;
 use crate::smb::dcerpc::*;
 use crate::smb::events::*;
